refactor: navigation and history into iextensionsIntegrated
