+++ src/extensions/default/DefaultExtensions.json
@@ -18,7 +18,6 @@
     "JavaScriptRefactoring",
     "JSHint",
     "MDNDocs",
-    "NavigationAndHistory",
     "Phoenix-prettier",
     "PrefsCodeHints",
     "QuickView",
