--- src/extensions/default/NavigationAndHistory/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "NavigationAndHistory",
-  "version": "1.0.0",
-  "dependencies": {
-  }
-}
