--- src/extensions/default/NavigationAndHistory/requirejs-config.json
@@ -1,1 +1,0 @@
-{}
