+++ src/extensionsIntegrated/NavigationAndHistory/html/recentfiles-template.html
