+++ src/extensionsIntegrated/NavigationAndHistory/html/recovery-template.html
