+++ src/extensionsIntegrated/NavigationAndHistory/keyboard.json
