+++ src/extensionsIntegrated/NavigationAndHistory/styles/recent-files.css
