+++ src/extensionsIntegrated/loader.js
@@ -38,4 +38,5 @@ define(function (require, exports, module) {
     require("./InAppNotifications/main");
     require("./NoDistractions/main");
     require("./Phoenix-live-preview/main");
+    require("./NavigationAndHistory/main");
 });
