+++ test/spec/Extn-RemoteFileAdapter-integ-test.js
